Alterações feitas nas entidades existentes
