+++ src/main/java/com/cuscuzcomjava/remotecontroller/entity/Producer.java
@@ -6,29 +6,5 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 
 @Entity
-public class Producer {
-
-    @Id
-    @GeneratedValue
-    @Column
-    private Long id;
-
-    @Column
-    private User user;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public User getUser() {
-        return user;
-    }
-
-    public void setUser(User user) {
-        this.user = user;
-    }
+public class Producer extends User{
 }
