+++ src/main/java/com/cuscuzcomjava/remotecontroller/entity/Reserve.java
@@ -1,11 +1,8 @@
 package com.cuscuzcomjava.remotecontroller.entity;
 
-import jdk.jfr.Enabled;
+import com.sun.istack.NotNull;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
+import javax.persistence.*;
 import java.time.Instant;
 
 @Entity
@@ -13,13 +10,15 @@ public class Reserve {
 
     @Id
     @GeneratedValue
-    @Column(name = "CL_ID")
+    @Column
     private Long id;
 
-    @Column(name = "CL_DATE_RESERVED")
+    @Column
+    @NotNull
     private Instant dateReserved;
 
-    @Column(name = "CL_ACTRESS_ID")
+    @ManyToOne
+    @JoinColumn
     private Actress actress;
 
     public Long getId() {
@@ -38,11 +37,11 @@ public class Reserve {
         this.dateReserved = dateReserved;
     }
 
-    public Actress getActress() {
-        return actress;
-    }
-
-    public void setActress(Actress actress) {
-        this.actress = actress;
-    }
+//    public Actress getActress() {
+//        return actress;
+//    }
+//
+//    public void setActress(Actress actress) {
+//        this.actress = actress;
+//    }
 }
