+++ src/main/java/com/cuscuzcomjava/remotecontroller/entity/User.java
@@ -1,20 +1,23 @@
 package com.cuscuzcomjava.remotecontroller.entity;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
+import javax.persistence.*;
 
-@Entity
+@MappedSuperclass
 public class User {
 
     @Id
     @GeneratedValue
-    @Column(name = "CL_ID")
+    @Column
     private Long id;
 
-    @Column(name = "CL_NAME")
+    @Column
     private  String name;
+
+    @Column
+    private String login;
+
+    @Column
+    private String password;
 
     public Long getId() {
         return id;
@@ -30,5 +33,21 @@ public class User {
 
     public void setName(String name) {
         this.name = name;
+    }
+
+    public String getLogin() {
+        return login;
+    }
+
+    public void setLogin(String login) {
+        this.login = login;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
     }
 }
